Refactoring of implementations complete
