--- Os/purge-me/Stubs/CMakeLists.txt
@@ -1,35 +1,0 @@
-####
-# F prime CMakeLists.txt:
-#
-# SOURCE_FILES: combined list of source and autocoding files
-# MOD_DEPS: (optional) module dependencies
-#
-####
-set(SOURCE_FILES
-  "${CMAKE_CURRENT_LIST_DIR}/Linux/FileStub.cpp"
-)
-set(MOD_DEPS
-  Fw/Cfg
-  Fw/Types
-)
-
-register_fprime_module()
-
-# Sets MODULE_NAME to unique name based on path
-get_module_name(${CMAKE_CURRENT_LIST_DIR})
-
-# Exclude test module from all build
-set_target_properties(
-  ${MODULE_NAME}
-  PROPERTIES
-  EXCLUDE_FROM_ALL TRUE
-)
-# Sets MODULE_NAME to unique name based on path
-get_module_name(${CMAKE_CURRENT_LIST_DIR})
-
-# Exclude test module from all build
-set_target_properties(
-  ${MODULE_NAME}
-  PROPERTIES
-  EXCLUDE_FROM_ALL TRUE
-)
