+++ Svc/ActiveLogger/CMakeLists.txt
@@ -13,9 +13,7 @@ set(SOURCE_FILES
 
 register_fprime_module()
 ### UTs ###
-set(UT_MOD_DEPS
-    Os_Stubs
-)
+choose_fprime_implementation(Os_File Os_File_PosixWithInterceptors)
 set(UT_SOURCE_FILES
   "${FPRIME_FRAMEWORK_PATH}/Svc/ActiveLogger/ActiveLogger.fpp"
   "${CMAKE_CURRENT_LIST_DIR}/test/ut/ActiveLoggerTester.cpp"
