+++ Svc/CmdSequencer/test/ut/MixedRelativeBase.cpp
@@ -10,7 +10,7 @@
 
 #include "Svc/CmdSequencer/test/ut/CommandBuffers.hpp"
 #include "Svc/CmdSequencer/test/ut/MixedRelativeBase.hpp"
-#include "Os/Stubs/FileStubs.hpp"
+#include "FileStubs.hpp"
 
 namespace Svc {
