+++ cmake/utilities.cmake
@@ -559,5 +559,4 @@ function(append_list_property NEW_ITEM)
     list(REMOVE_DUPLICATES LOCAL_COPY)
     set_property(${ARGN} "${LOCAL_COPY}")
     get_property(LOCAL_COPY2 ${ARGN})
-    message(STATUS ">>>>>>>>> Appended: ${LOCAL_COPY2}")
 endfunction()
